+++ forge/src/main/java/com/lx862/jcm/mapping/LoaderImplClient.java
@@ -1,8 +1,12 @@
 package com.lx862.jcm.mapping;
 
 import com.lx862.jcm.mixin.minecraft.NativeImageAccessor;
+import it.unimi.dsi.fastutil.objects.ObjectArrayList;
+import net.minecraft.client.player.AbstractClientPlayer;
 import org.mtr.mapping.holder.*;
 import org.mtr.mapping.mapper.ScreenExtension;
+
+import java.util.List;
 
 /**
  * Forge implementation via Mojang mapping
@@ -23,5 +27,13 @@ public class LoaderImplClient {
 
     public static long getNativeImagePointer(NativeImage nativeImage) {
         return ((NativeImageAccessor)(Object)nativeImage.data).getPixels();
+    }
+
+    public static List<PlayerEntity> getWorldPlayers(ClientWorld world) {
+        List<PlayerEntity> players = new ObjectArrayList<>();
+        for(AbstractClientPlayer entity : world.data.players()) {
+            players.add(new PlayerEntity(entity));
+        }
+        return players;
     }
 }